feat: `/register`で作成した商品登録フォームを表示するよう変更
